create login and registration system

+++ LibraryProject/src/main/java/com/example/libraryproject/model/entity/User.java
@@ -52,7 +52,7 @@ public class User {
 
     @Column(name = "review_count")
     @NonNull
-    private int reviewCount;
+    private Long reviewCount;
 
     public User() {}
